fix or suppress MSVC warnings

+++ chips/clk.h
@@ -125,7 +125,7 @@ void clk_init(clk_t* clk, uint32_t freq_hz) {
 
 /* freq_hz being 64 bit is not a bug, needed to prevent a 32-bit overflow */
 uint32_t clk_us_to_ticks(uint64_t freq_hz, uint32_t micro_seconds) {
-    return (freq_hz * micro_seconds) / 1000000;
+    return (uint32_t) ((freq_hz * micro_seconds) / 1000000);
 }
 
 uint32_t clk_ticks_to_run(clk_t* clk, uint32_t micro_seconds) {
